Update API of py::oiter
