+++ c/frame/__init__.cc
@@ -25,7 +25,6 @@
 #include <vector>
 #include "python/_all.h"
 #include "python/list.h"
-#include "python/oiter.h"
 #include "python/oset.h"
 #include "python/string.h"
 #include "utils/alloc.h"
